zen_web/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-queue-manager==4.0.0
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.10
Django==5.1.7
django-allauth==65.6.0
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-environ==0.12.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.1
packaging==25.0
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
pymemcache==4.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
zope.event==5.0
zope.interface==7.2
